+++ flowgen/app/src/flow.rs
@@ -184,12 +184,12 @@ pub enum Error {
     /// Error in Braze export user IDs task.
     #[error(transparent)]
     BrazeExportUsersIds(#[from] flowgen_braze::export::users::processor::Error),
-    /// Error in Mongo collection task.
+    /// Error in MongoDB collection task.
     #[error(transparent)]
-    MongoCollection(#[from] flowgen_mongo::collection::Error),
-    /// Error in Mongo change stream task.
+    MongoDbCollection(#[from] flowgen_mongodb::collection::Error),
+    /// Error in MongoDB change stream task.
     #[error(transparent)]
-    MongoChangeStream(#[from] flowgen_mongo::change_stream::Error),
+    MongoDbChangeStream(#[from] flowgen_mongodb::change_stream::Error),
     /// Failed to store background task handles for later monitoring.
     #[error("Error storing background task handles")]
     BackgroundHandlesStoreFailed,
@@ -1946,11 +1946,11 @@ async fn spawn_task(
             )
         }
 
-        TaskType::mongo_collection(config) => {
+        TaskType::mongodb_collection(config) => {
             let config = Arc::new(config);
             tokio::spawn(
                 async move {
-                    let mut builder = flowgen_mongo::collection::ProcessorBuilder::new()
+                    let mut builder = flowgen_mongodb::collection::ProcessorBuilder::new()
                         .config(config)
                         .task_id(task_id)
                         .task_type(task_type_str)
@@ -1968,12 +1968,12 @@ async fn spawn_task(
             )
         }
 
-        TaskType::mongo_change_stream(config) => {
+        TaskType::mongodb_change_stream(config) => {
             let config = Arc::new(config);
             tokio::spawn(
                 async move {
                     let mut builder =
-                        flowgen_mongo::change_stream::ChangeStreamReaderBuilder::new()
+                        flowgen_mongodb::change_stream::ChangeStreamReaderBuilder::new()
                             .config(config)
                             .task_id(task_id)
                             .task_type(task_type_str)